fix bug when ability is deleted while another is being edited
When editing an ability, we cache the original to restore it if the user cancels. The cached original is given an id value equal to the index of the ability being edited in the abilities list. If another ability with a lower index is deleted, the id value of the cached original no longer matches the index of the ability being edited. This causes issues when the user cancels or saves their edit and we attempt to restore from or clear the cache. This commit fixes this issue.