sort abilities by cost, low-to-high
This required decoupling ability caching from the index of the ability being edited. The cached ability is now linked to its original by uuid, which is immutable and therefore less prone to errors when abilities are re-ordered. Only delete and update use index of the ability being edited. These actions are safe as they are effectively instantaneous, so ordering cannot change during delete or during update.